リファクタリング：画像に割り当てられたタグ取得処理

+++ Rialto/Models/DataModel/ImageInfo.cs
@@ -13,7 +13,7 @@ namespace Rialto.Model.DataModel
         {
         }
 
-        public long ImgID { get; set; }
+        public int ImgID { get; set; }
 
         private BitmapImage _DispImage = null;
